+++ frontend/src/hooks/domain/ui/useScrollUITriggers.ts
